requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0
googlemaps==4.10.0